feat: session로그인 기능 추가
SessionManager 클래스 추가 및 유저 컨트롤러 수정

+++ src/main/java/com/gdsc/festivalholic/controller/LikesController.java
@@ -1,14 +1,19 @@
 package com.gdsc.festivalholic.controller;
 
+import com.gdsc.festivalholic.SessionManager;
+import com.gdsc.festivalholic.config.exception.ApiException;
+import com.gdsc.festivalholic.config.exception.ErrorCode;
 import com.gdsc.festivalholic.config.response.ResponseDto;
 import com.gdsc.festivalholic.config.response.ResponseUtil;
 import com.gdsc.festivalholic.controller.dto.likes.LikesRequestDto;
+import com.gdsc.festivalholic.domain.users.Users;
 import com.gdsc.festivalholic.service.LikesService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -32,16 +37,29 @@ import org.springframework.web.bind.annotation.RestController;
 public class LikesController {
 
     private final LikesService likesService;
+    private final SessionManager sessionManager;
 
     @Operation(summary = "좋아요", description = "")
     @PostMapping("")
-    public ResponseDto<Long> like(@RequestBody LikesRequestDto likeRequestDto) {
+    public ResponseDto<Long> like(@RequestBody LikesRequestDto likeRequestDto, HttpServletRequest httpServletRequest) {
+        Users users = (Users) sessionManager.getSession(httpServletRequest);
+
+        if(users == null) {
+            throw new ApiException(ErrorCode.NOT_LOGIN);
+        }
+
         return ResponseUtil.SUCCESS("좋아요 성공", likesService.insert(likeRequestDto));
     }
 
     @Operation(summary = "좋아요 취소", description = "좋아요 취소 API")
     @DeleteMapping("/{likesId}")
-    public ResponseDto delete(@Parameter(description = "좋아요 인덱스 번호") @PathVariable Long likesId) {
+    public ResponseDto delete(@Parameter(description = "좋아요 인덱스 번호") @PathVariable Long likesId, HttpServletRequest httpServletRequest) {
+        Users users = (Users) sessionManager.getSession(httpServletRequest);
+
+        if(users == null) {
+            throw new ApiException(ErrorCode.NOT_LOGIN);
+        }
+
         likesService.delete(likesId);
         return ResponseUtil.SUCCESS("삭제 완료", likesId);
     }
